working on the methods property
using v-model directive for two way binding. Increment amount model by default is set to 5 in data property and can be changed in the input field. The count can be incremented based on the incrementAmount model. Once the button clicked, it increments the count with the set amount by incrementAmount model